refactor:extract ThresholdUtils remove featureEnvy
